requirements.txt
PyQt5==5.15.9
opencv-python==4.8.0.74
opencv-contrib-python==4.8.0.74
numpy==1.24.3
google-generativeai==0.3.1
pillow==10.0.0
sqlite3-api>=2.0.0
scenedetect==0.6.2
openai==0.27.8
